Start process of making 'low'/'high' return aligned bounds by default
This is a start at a process that was proposed awhile back to make
'.low' and '.high' queries on ranges and domains return the aligned
bounds rather than the pure/potentially unaligned ones.  For instance,
for '1..10 by 2' the pure high bound is 10 while the aligned high
bound is 9.  Historically, '.high' would return the pure bound and
'.alignedHigh' would return the aligned one, but we've been discussing
changing '.high' to the aligned one since that's what people tend to
want more often (and those who want the pure one will probably be
expert enough to know how to ask for it).

To that end, this PR introduces '.lowBound' and '.highBound' as ways
to query the pure bounds and introduces a new warning when applying
'.low'/'.high' to a strided range indicating that the behavior is
changing and telling people how they can sign up now.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>