Deprecate OS.POSIX symbols defined in SysBasic and move docs over
This deprecates the symbols that were duplicates between SysBasic and
OS.POSIX (from the former) and moves their documentation to the
latter.  It also updates standard modules to make use of OS.POSIX
and/or to pick and choose the specific symbols it needs from SysBasic
to avoid deprecation warnings or duplicate symbols.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>
Signed-off-by: Brandon Neth <nethbrandone@gmail.com>