Update test/library/standard/Math/promotion directory for new Math organization
The tests in the promotion directory are generated by a script in the
`generate_tests` subdirectory.  This script is not run by start_test, it has to
be run by hand.  Adjusted the script to add a use for the symbols that now
require an explicit use, by making a separate category for them and adjusting
the output.  Also adds the updated test file generated by the script

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>