Update test and its helper files to continue to exercise the same behavior
When Math became not used by default, foo2.chpl began to resolve (which it
should do).  However, the test was added to exercise the behavior when naming a
module with the same name as a standard module that is included by default, so
the appropriate choice for this test was to rename the module it was trying to
include such that it conflicts with AutoMath (which is included by default).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>