Quick patch to mac specific bug introduced in previous PR
This PR: https://github.com/chapel-lang/chapel/pull/19851 was Linux specific
(it searches for a dynamic lib and assumes the extension is .so). I generalized
it to also work on mac and windows systems.  This is still not ideal (maybe
there's another file extension I'm forgetting about?). But it's better.

---
Signed-off-by: Andy Stone <stonea@users.noreply.github.com>
Signed-off-by: Brandon Neth <nethbrandone@gmail.com>