Add future for tertiary method namespace / visibility leak issue
This adds a future that demonstrates that, today, when we define a
tertiary method, it makes other symbols that are visible within the
original type's scope visible to the scope in which the tertiary
method is defined.  We (Lydia, Michael, me) think that this is an
oversight and not intended.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>