Remove last uses of Math from AutoMath by returning the body of log1p
While a deprecated version of log1p still lives in AutoMath, put its
functionality in that module in an undocumented `chpl_` function and have the
version we intend users to call in Math call out to that function.  This means
that programs without an explicit use of the Math module won't have to compile
it and its contents.

Also make the deprecated versions of logBasePow2 call the `chpl_` version of
that function.  Unlike the `chpl_log1p` definitions, `chpl_logBasePow2` will
stay once the deprecated versions have been removed, due to `log2`'s reliance on
it.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>
Signed-off-by: Brandon Neth <nethbrandone@gmail.com>