Comment out recursive primitive call
Engin caught that a case of `.lowBound` should've been `alignedLow`,
but when I wrote a test to exercise this and make sure it works, I
found that it didn't work even without the change due to the compiler
balking at an infinite recursion.  Here, I'm commenting out the
primitive in order to avoid the infinite recursion and adding a test
to show that the call can now be made.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>