Fix installation of mason
Historically, mason was built into `tools/mason`, so that's where
our 'make install' command would try to install it from.  However,
in #16154, it was changed to build directly into `bin/[whatever]/`
causing `make install` to stop finding and copying it (except for
those of us who still had an ancient version of `tools/mason/mason`
lying around).

Here, I'm updating the make install script (`install.sh`) to copy
it from the correct location, and also adding a 'make mason' step
to our test of the installation logic, followed by a check to see
that it got put where expected to try and prevent this from happening
again in the future.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>