Remove SysCTypes and [s]size_t types (deprecated in 1.26.0)
This is a follow-up to #19284 that removes the size_t and ssize_t
type aliases, along with the SysCTypes.  The former were replaced
by 'c_[s]size_t' types while the latter was replaced by 'CTypes'
(as far as users are concerned) and 'ChapelSysCTypes' (internally).

In doing this, I found a few places that were still referring to
SysCTypes incorrectly, most notably the submitted pidigits4.chpl
version.  Here, I took the approach of putting a temporary
SysCTypes.chpl module into that directory so that we could continue
getting correctness and performance tests for it going forward.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>