Update some tests to throw --memTrack
These tests were relying on the old behavior of getting some logging
when optimizations weren't enabled. This happened to work for
test-memLog because we only ran it in a std gasnet config, and
allocationCounts was skipping for `--fast` and trying to prediff some
things out for our comm none config that throws memtrack.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>