Avoid duplication of qualified function deprecation warnings
All functions were generating deprecation warnings already when the final symbol
being called was determined.  Additionally, if a symbol found with the name
was deprecated, this check could fire, even if an undeprecated candidate
ultimately ended up being the final choice.  So skip the warning at scope
resolution if the symbol matched was a function, trusting function resolution
to handle it (as it does today)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>