Disable readline() for strided arrays for now
Though I got it working for positively strided arrays where it wasn't
before, BHarsh pointed out that it wasn't working correctly for
negatively strided arrays.  Meanwhile, while this PR has been in
limbo, Shreyas is working on a fairly significant rewrite of this
routine, and I don't want him to have to worry about that case.
Ultimately, I think we should support strided arrays, though (in the
"support as many operations on as rich a set of arrays as possible"
sense).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>