Partial packaging for Arch User Repository (AUR)
Adds packaging builds for Arch (Linux) User Repository (AUR).
The chapel package for AUR tracks the static 1.26 release, but is non-functional due to issues in 1.26 that cause build issues on Arch. These issues have been resolved, but the chapel AUR package will be left broken until 1.27 releases (the online description points to chapel-git).
The chapel-git package for AUR tracks the main branch of the chapel repo, and is fully functional.
Future work includes:
- updating chapel AUR to use 1.27 upon its release
- fixing the git describe behavior for the main branch (currently main reports as version 1.24 via git describe, while it is really 1.26 at the time of writing).

Signed-off-by: ghbrown <gabriel.h.brown@gmail.com>