Remote cache instrumentation for prefetches and read aheads
This commit contains 6 new counters for CommDiagnostics that
relate specifically to the remote cache. The counters will indicate
how many prefetches and read aheads were performed, and of those, how
many went unused before being evicted and how many we had to wait for
because they were still in-flight on the network when we tried to access
the data.

Names of the counters are open to discussion. There is one issue that is
present, which is a fixed number of inaccurate counts whenever the
diagnostics are turned off. Whatever is left in the remote cache is
not accounted for.

Signed-off-by: Thomas Rolinger <tbrolin@cs.umd.edu>