Fix use-after free in scope resolve cache
For the test array-initialization-patterns-dists.chpl, I observed
a use-after-free within scope resolution.

An earlier commit changed this in ChapelBase:

  inline operator ==(a: int(64), b: uint(64))

to

  inline operator ==(a: int(?w), b: uint(w))

That function '==' is one of the functions that buildDefaultFunctions
looks for. At the same time, 'normalize' deletes this function
and replaces it with other functions in 'cloneParameterizedPrimitive'.
The simple fix employed here is to clear the scope resolution
cache after scope resolve proper runs.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>