Keep "called" modules alive for the sake of generating errors for them
Prior to this PR, a module that was "called" but otherwise dead was
dead-code eliminated, causing the compiler error that we'd generate
further along in compilation to refer to dead code.  This PR keeps
such called modules alive for the sake of generating a legal error.
It seems that we could also generate the error here, but this
approach has the benefit of emitting all instances of this error
in a single place.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>