Trim one yield off of the unbounded integer ranges
It turns out that the additional yield, while very nice in terms
of permitting the full range of integers to be yielded, has the
downside of making the loops thwart our zippered range iterator
optimization in cases like 'zip(whatever, 1..)'.  This commit
removes the extra 'yield' for integer ranges, and rewrites the
bounds for the bool and enum cases so that they still yield the
full range of values.

Note that if we had distinct standalone and follower serial
iterators, we could use the extra yield in the standalone case
(where it's arguably more important) and have the follower case
not contain it.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>