Resolve param for loops
Resolves param loops by creating a new Resolver for each iteration, and
storing the collected ResolutionResultByPostorderIDs in a
ResolvedParamLoop result. Each iteration stores a different param
resolution result for the AST representing the index variable.

These temporary Resolvers don't inherit their parent's scopeStacks or
their ResolutionResultByPostorderID values directly. Instead, a
'parentResolver' reference is stored in the Resolver type and used by
'typeForId' to determine whether a parent resolver 'owns' the resolution
results for the desired scope.

The ResolvedParamLoop is finally stored in the ResolvedExpression for
the For loop.

Remaining work:
- handle strided ranges
- handle ranges with '#' or '<' operators

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>