First draft of a ResolvedVisitor
This is a first draft of a ResolvedVisitor type, which is meant to
handle iterating over the AST and making resolution results available to
the client type.

Clients of this type will implement their own visitors, e.g.
MyVisitor, which will implement enter/exit methods that accept two
arguments:
- An AST node
- A reference to a ResolvedVisitor<MyVisitor>

The ResolvedVisitor stores a ResolutionResultByPostorderID that will
contain the relevant results for the corresponding AST.

The ResolvedVisitor will handle param for loops by first invoking the
client 'enter' method to determine whether it should recurse. If so, it
will traverse once over the iterand, and then will traverse for each
iteration of the loop body with the corresponding results from the
ResolvedParamLoop.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>