Explore no longer merging real/imag/complex in heuristic
Need to enable commented-out real/imag param fns from
PR #12386.

Interestingly, before & after this change,
param real(32) + param complex(128) is complex(64)

  param paramReal32: real(32) = 1.0;
  param paramComplex128: complex(128) = 1.0 + 2.0i;

  param z = paramReal32 + paramComplex128;

  writeln("z = ", z, " : ", z.type:string);

Anyway, with this change, param arguments passed to f
in allNumericsBinaryParamParam.chpl end up with
complex(128) instead of complex(64). That is because
the complex(64) overload is filtered out (since the
type complex(64) is not mentioned at the call site).

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>