Fix prediff to remove only logging text
This test logs all array accessor calls, which could happen in any
number of places in the internal modules. If those array accessor calls
are shuffled around with IO changes, then the output of this test could
change unexpectedly and in a confusing way (e.g. an extra space
appearing, or disappearing).

This prediff targets and removes the exact array logging text, leaving
the other output untouched.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>