Make slurm-srun `--mem` overridable
In 13667, we started throwing `--mem=0` for the slurm-srun launcher to
get access to all memory on each node. This is still the default we
want, but a secondary use for `--mem` is to request the high memory
nodes (e.g. to select 512G nodes in a partition that has mixed 256/512G
ones.) Update the code to allow overriding with `CHPL_LAUNCHER_MEM`.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>