Suppress unbounded enum/bool tests for --baseline; add future showing bug
When #19991 was merged, we started seeing failures for the new tests
with --baseline testing where the loops would go into an infinite
loop.  This was no worse than the behavior prior to #19991 (also an
infinite loop), and turns out to be a pre-existing condition (at least
as far back as 1.25.0) that seems to plague iterator methods of a
certain kind.  In this PR, I've illustrated this with
no-inline-method-iter.chpl which shows an iterator that works by
default but not with --no-inline-iterators.

For this reason, I've added a '.suppressif' to the tests that were
causing infinite loops in testing to silence them when compiled with
--baseline or --no-inline-iterators, and refactored the tests a bit to
detect the infinite loop early rather than waiting for a timeout and
generating too much data.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>