Stop trying to specialize for unbounded iteration
Prior to this change, the compiler assumed that any method invoked
on an unbounded range was infinite, but of course this isn't
necessarily the case.  E.g.,

iter range.yieldOnce() {
  yield 1;
}

is not infinite even if called on `1..`.  In this PR, we stop
special-casing unbounded ranges like this, believing it may
make the compiler only more correct.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>