Restore Omni-Path documentation
The latest gasnet version restored support for running on Omni-Path with
the ofi conduit. Add documentation on how to run on this network. These
docs are somewhat based on the InfiniBand documentation. Instructions
have been testing on Grid5000 and Sandia Omni-Path systems.

Note that this is for the EOL Intel Omni-Path Architecture (OPA) and not
the upcoming Cornelis Omni-Path Express (OPX).

Resolves 13100

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>