Suppress ranges.chpl primer for --baseline
As proof that I'm a slow learner, this is the same pre-existing bug /
timeout that I suppressed for other unbounded range loop tests in
this suppresses the failures for --baseline and --no-inline-iterators.
Unlike there, I didn't change the test to make it give up earlier
(since it's a primer), but did set the timeout lower to make it less
of a resource hog.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>