Reduce comm for storing block dist prescan results
Previously, we would PUT our outputSync class and prescan result to
locale 0 and then do an implicit on-stmt as part of the sync to mark
that the prescan result is ready.

This modifies the code to make the on-stmt explicit and does the
outputSync and prescan assignment inside it. This removes the PUTs (and
the GETs to acquire DR metadata). The values are now sent as part of the
on-stmt bundle (which required making some things const to enable RVF.)

This reduces comm for just the scan (excluding the result array
creation) on non-0 locales from 10 GETs and 2 PUTs down to just 6 GETs.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>