Write output value when updating output ready sync in block dist scan
Instead of GET'ing the value after reading our sync, just set the value
while we're updating the sync. This updates the BoxedSync to keep a
value stored with it. It's not generally safe because it splits the
value reading/writing from the FEB manipulation, but it works for the
single reader/writer case here.

Note that I also switched from using `sync bool` to `sync int` because
we support native qthread sync vars for int and not bool. This shouldn't
matter too much, but seemed worthwhile while here. (See 7532 for
supporting more types for native sync vars.)

This reduces comm for the non-array creation part of scans from 3 GETs
on non-0 locales down to 2 (note that the 2 remaining GETs are because
our arrays in this coforall yield references that can't be RVF'd.)

I'll also note that I think we could use `single` here instead of `sync`
but I didn't do that because we don't support using qthreads for that.
And lastly you could use the sync for the storage too for result types
that are supported by syncs, but I don't expect any real performance
benefit from that, so I didn't think it was worthwhile to do now.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>