Added test for remote cache readahead CommDiagnostics counters
These counters are not easily tested when compared to the prefetch
counters, since readaheads are performed automatically versus the
manually inserted prefetches. To this end, the added test ensures that
when non-adjacent GETs are performed, we do not see any readaheads. It
also tests that when we do adjacent GETs, the number of readaheads is
non-zero and there are no unused readaheads. The exact number of
readaheads performed is not something I know how to determine as a
function of the number of GETs performed. Furthermore, the number
of readheads that were waited on cannot be easily determined since
it can be influenced by the system's interconnect latency.

Also removed the Time module use-statement from the prefetchCounters.chpl
test, since it is not used.

Signed-off-by: Thomas Rolinger <tbrolin@cs.umd.edu>