Bring block dist scan performance on par with local DR arrays
Optimize block dist scans by passing in local views of the result array
to preScan/postScan. This means the indexing in these methods will be on
default rectangular arrays instead of block and DR indexing is cheaper.
Note that this is only legal when the array being scanned and the result
array have the same distribution so reuse the fast-follower checks to
ensure that's the case for this optimization to trigger.

This brings the performance of block scans on par with DR. On a 128-core
rome node I see a 128GB scan go from 2.35s to 2.25, which matches the
performance of DR.

We're operating on wide DR arrays for block, but that doesn't seem to
impact performance at least when using all cores since we're bandwidth
bound. When using fewer cores there is still a difference however.

Part of Cray/chapel-private 3505

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>