Clean up C code generation of chained if-then-else statements
This does a spot-check in C code generation for chained if-then-else
blocks so that rather than generating:

```c
if (...) {
} else {
  if (...) {
  } else {
    if (...) {
    } else {
    }
  }
}
```

we generate:

```c
if (...) {
} else if (...) {
} else if (...) {
} else {
}
```

Essentially, it checks to see whether there's only one statement in
the else block and whether that one statement is itself a conditional.
If so, it codegens it directly; otherwise it does the normal thing.

Due to the compiler's insertion of temps, cases that _seem_ like they
should generate clean chained conditionals don't always, such as
test/parsing/bigifthenelse.chpl.  Perhaps there's more we could do to
denormalize to address such cases, but in the meantime, this seems
like a step forward.

Resolves #3519

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>