Fix `CHPL_COMM_USE_(G|LL)DB` on mac
In 8aca2b3b575 we added a `find_executable` helper, which changed a
`which` call to `/usr/bin/which --skip-alias --skip-functions`. These
options ended up not being portable to the `which` command on mac, which
meant features like `CHPL_COMM_USE_LLDB` couldn't find executables they
needed and weren't working as a result. To fix this just go back to
using a plain `which`. The motivation for using the skip options was:

> In case the user has an alias or shell function with different
> behavior. I think we want to be using "classic which" rather than
> anything derivative.

But I don't think this gave us trouble in the past.

Part of Cray/chapel-private 3374

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>