Add EnumParam subclass to Param.
An enum literal like `BoundedRangeType.boudned` is a param. Prior to this
commit, dyno didn't have a way to represent a Param that isn't a number or
string. Adding an `EnumParam` class makes it possible to do so. The new class
contains a reference to the AST node corresponding to the enum element.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>