Re-enable counts (#) on ranges to take bool values
In #20098, I unwittingly disabled support for applying boolean count
values to ranges because 'integral' arguments don't accept booleans.
It turns out Arkouda relied on this (our testing system didn't), so
here I'm re-enabling it for the time being as it wasn't my intention
to change the behavior.

For me, there are a few questions here:

* Should we continue to support boolean counts?  (the alternative
  being to require the user to cast them to an integer)

* Should integral formal types accept boolean actuals?  Is it
  inconsistent that our previous implementation that accepted
  int(?w)/uint(?w) counts accepted bools when integral doesn't?

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>