Disallow duplicate targetLocales for Block arrays
Block didn't explicitly disallow duplicate targetLocales before, but
several operations didn't work correctly and some upcoming optimizations
we want to make would further break things. This PR changes that to
explicitly disallow duplicates by default, and adds an undocumented
allowDuplicateTargetLocales param in case we or users really want to
play with targetLocale oversubscription. When this param is thrown, we
also workaround the known bugs at the expense of performance. I suspect
if there is any demand for this we'd eventually make a new distribution,
but for now we wanted some way to get the old behavior.

Note that this is an always on check instead of only on with bound
checks or something. We could make it a bound check only, but it should
be trivially fast in the context of distributed array creation.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>