Defuturize: ref array formal can't accept iterable
This PR resolves three futures by asserting the following semantics.
A formal argument of an array type cannot accept an iterable, such as
a promoted expression, when the formal's intent is `ref` or `const ref`.
This includes the default and `const` abstract intents.

The rationale is that a `ref` formal requires an actual argument of the
same type. An iterable expression does not have an array type and there is
no underlying array to refernce.

To force creation of an array out of an iterable when passing the latter
to a formal argument, the formal argument must have an `in` or `const in`
intent, see `test/users/thom/arrArgIn-promote.chpl` .

Next steps: should a `ref` formal accept an array returned by a function
with  the default return intent? What if it is a `const ref` formal and/or
the function's return intent is `const` ? For example:

```chpl
proc writeArr(ref X) {  // what if X has a `const ref` intent?
  writeln(X);
}

proc addEm() {    // what if addEm() has a `const` return intent?
  return [1,2];
}

writeArr(addEm());
```

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>