Skip test_10k task tests that communicate under fifo
gasnet only allows so many threads to communicate concurrently (256 by
default) and this test can result in 10,000 threads. We haven't seen
this test fail previously, but it does on m1 macs now. I'm not sure if
time-slicing on m1 is different or something is slower or faster but
somehow more concurrent threads are calling into gasnet.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>