Make a string start on the same line as the opening quote
The string that defines a bash script in this test started with a newline
character because it was written as:

const script = """
<script contents>
""";

This recently stopped working on some test systems because there was a blank
line before the #! line in the script. Move the <script contents> to the same
line as the opening triple quote.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>