Add two tests of fcfs as fields
One just checks saving a fcf in a field and calling it in a method.  The other
is the case I encountered when trying to implement distributed maps, where I
was attempting to use a `nothing` type in the case where I was relying on a
default hasher, which triggered a resolution failure.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>