Update future to reflect correction from Brad and Michael and add another test
The other test is what I should have written (and works as intended).  The old
future has now been updated to be an "error message" future, since it should not
work but it should generate a better message to the user.  I updated the .good
file with a potential error message, but we may not generate that exact one when
the issue is resolved

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>