Detect unknown type even when it does not have a Type*
We avoid resolving functions when any of their actuals
has an unknown type. Currently, this is implemented by
checking if the QualifiedType's Type*, if any, is
unknown. However, plenty of other places in the code
produce empty QualifiedType, with kind UNKNOWN and
no Type*. These slip by this check, and cause
assertion errors when functions called with
unknown-typed actuals are instantiated. Instead,
we should use QualifiedType::isUnknown.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>