Add env var to force skip tests that use ZMQ
Our tests that use ZMQ assume that if a shared library can be found then
so can the headers and a static lib, but this isn't always the case.  On
many Cray systems just `libzmq.so` is available but nothing else.

Longer term, we probably want more of a configure-like check for zmq
functionality, but for now just add an option to tell the skipifs that
not all zmq functionality is available.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>