Adjust how the production compiler reports dyno errors...
Adjust the production compiler to store errors and then "realize"
them at a later point. This allows the compiler to call queries
to inspect the errors while avoiding the possibility of a cycle (as
could happen if queries were called in the error handler).

Have production print a header for dyno errors in the same way that
it does for old AST errors. This header lists the enclosing function
or module name.

Fix bugs in the checks for nested decorators in new expressions. Add
a check for nested decorators in type constructor calls.

Adjust parse checks test to use 'parseFileToBuilderResult' instead
of using the parser directly, as the latter does not store file
paths in the context.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>