Use narrow arrays for blockdist scan
Now that duplicate targetLocales are no longer allowed for block arrays,
we know that the local array pointer `myLocArr` is safe to use. Using it
instead of `locArr[locid]` allows us to do scan operations on narrow
pointers, which reduces any wide pointer overhead.

On a 128-core Rome CPU running with just 2 cores this takes a 16G Block
scan from 4.9s to 2.9s, bringing it on par with DR. This has no real
impact when using all 128 cores since we're bandwidth bound. I expect
this to improve performance on systems with fewer cores and in cases
where not all cores are used (e.g. serial scan in a parallel region)

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>