Copy some default and const intents from production compiler
The production compiler specifies default intents for a lot
of the built-in types. However, Dyno does not yet know how
to handle intents for these types. This results in assertion
errors whenever a function is called that uses any such built-in
types.

For the following test program:

    proc test(s: c_string) {}
    test("hello".c_str());

Currently, the compiler reports an assertion error:

    Assertion failed: (t->genericity() != Type::CONCRETE)

With this change, however, the call is correctly resolved.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>