Switch away from tree-based privatization
Switch from a manual tree-based privatization scheme to our normal
`coforall loc in Locales do on loc`. The tree-based scheme was added way
back in 8e63e2ba19, likely before coforall+on was optimized.

A tree-based spawn can certainly be faster than the one-to-many spawn
currently used by the coforall+on pattern, but coforall+on is what we
use everywhere else and if it becomes a scaling bottleneck I think we
want to make that tree-based instead of having point solutions for
things like privatization.

I don't have access to larger machines to run on at the moment, but some
initial results showed better performance at 256 nodes of an XC and I
see a ~2x speedup for creating a trivial block domain and array on 16
nodes of an XC and CS:

```
./distCreate --size=arraySize.tiny --dist=distType.block -nl 16
```

16-node-cs-hdr

| config | domInit  | arrInit  |
| ------ | -------: | -------: |
| before | 0.00731s | 0.00106s |
| after  | 0.00464s | 0.00069s |

16-node-xc:

| config | domInit  | arrInit  |
| ------ | -------: | -------: |
| before | 0.00110s | 0.00022s |
| after  | 0.00055s | 0.00013s |

Resolves Cray/chapel-private 503
Part of 20197

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>