Optimize block dist privatization
Use a record instead of a tuple to shared privatzed dist information.
Allows some fields to be RVFd, and allows us to RVF the array record
wrappers for the targetLocales and locDist arrays. (we do this array
record wrapper trick for domain and array privatization already.) This
takes us from 20 GETs to 10 GETS.

We could probably get this down to a GET per array if we RVFd the
domains and then somehow passed the domains to the functions that end up
doing bulk transfers, but I'm not sure how much work that would be.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>