Disable block dist privatization when `--no-privatization` it thrown
This makes it easier to test whether comm is coming from the actual
dist/dom/arr creation or privatization. There's probably a bug higher in
the stack that's trying to privatize when it shouldn't be. That should
probably be fixed, but for now this is an easier way of testing.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>