Remove a compiler nondeterminism from PR #19561
PR #19561 changed dead code elimination in a way that introduces
nondeterminism. It changed a set to a SmallPtrSet but that set is
traversed and added to a work queue. The result is, with the SmallPtrSet,
the order of elements in the work queue depends on what addresses the
allocator got, which can vary run-to-run with ASLR.

We noticed this issue because there were nightly failures with these
tests in gasnet-asan testing:

 distributions/bharshbarg/stencil/rankChange.chpl
 distributions/robust/arithmetic/collapsing/test_domain_rank_change1.chpl
 distributions/robust/arithmetic/collapsing/test_domain_rank_change_OOB.chpl

This PR reverts just the change in PR #19561 to dead code elimination to
remove that nondeterminism.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>