Add 3 distinct mason modes and `mason modules` command This PR adds a conception of 3 distinct mason "modes": (1) a mason "package" - intended to run as a standalone     program, has a main function, fully featured (2) a mason "library" - intended to be `use`d in other     user code, does not have a main function, can't be     run standalone (3) a mason "light" project - just creates a toml file     that can track dependencies and generate the flags     to use those dependencies in any `chpl` command
Also, a new command `mason modules` is added that will
return a string of the include flags that would be added
onto a `chpl` command to use the dependencies specified
in a `Mason.toml` file that enables a little bit more
flexibility into building programs that use mason deps.

Docs are not included in this PR as there isn't a natural
fit for this in the existing mason docs, but the docs will
be undergoing an overhaul in the near future and will be
added then (see https://github.com/Cray/chapel-private/issues/3599).

Signed-off-by: Ben McDonald <mcdobe100@gmail.com>